+++ packages/views/issues/components/index.ts
@@ -7,3 +7,4 @@ export { CommentCard } from "./comment-card";
 export { CommentInput } from "./comment-input";
 export { ReplyInput } from "./reply-input";
 export { IssueMentionCard } from "./issue-mention-card";
+export { IssueChip } from "./issue-chip";
